Captura de error CSRF

+++ database/migrations/2021_07_29_164017_create_table_producto.php
@@ -17,7 +17,7 @@ class CreateTableProducto extends Migration
             $table->bigIncrements('id');
             $table->string('nombre');
             $table->text('descripcion')->nullable();
-            $table->integer('categoria_id')->unsigned();
+            $table->bigInteger('categoria_id')->unsigned();
             $table->integer('stock')->default(0)->unsigned();
             $table->decimal('precio_promedio')->default(0);
             $table->decimal('porcentaje_ganancia');
